add puzzle game score to backend, fixing backend routing
