+++ client/src/context/auth.jsx
@@ -47,7 +47,7 @@ const AuthProvider = ({children}) => {
     };
 
     return (
-        <AuthContext.Provider value={{user, login, logout}}>
+        <AuthContext.Provider value={{user, login, logout, backendUrl}}>
             {children}
         </AuthContext.Provider>
     );
